+++ lib/typecheck/typecheck.mli
@@ -269,6 +269,8 @@ type env = {
   no_panic_mod : bool;
   no_panic_modules : string list;
   nonexhaustive_match_spans : Ast.span list ref;
+  linear_ok_ids : (int, unit) Hashtbl.t;
+  linear_generic_uses : (Ast.span, string * int list * ty list * ty) Hashtbl.t;
   cap_producer_ivars : (int, Ast.span) Hashtbl.t;
   cap_narrow_factory_fns : (string, Ast.span) Hashtbl.t;
   cap_dicts : (string * string) list;
@@ -299,7 +301,7 @@ val make_env : Err.ctx -> (Ast.span, ty) Hashtbl.t -> env
 val lookup_ctor : StrMap.key -> env -> ctor_info option
 val add_ctor :
   string -> ctor_info -> ctor_info list StrMap.t -> ctor_info list StrMap.t
-val instantiate : ?use_span:Ast.span -> int -> env -> scheme -> ty
+val instantiate : ?use_span:Ast.span -> ?use_name:string -> int -> env -> scheme -> ty
 val stdlib_source_files : string list ref
 val cap_strict_ceiling : bool ref
 val builtin_cap_table : (string * string) list